requirements.txt
requests
pandas
google-cloud-bigquery
db-dtypes